Implement misc api endpoint

+++ extension.php
@@ -1,37 +1,46 @@
 <?php
 
+require_once(__DIR__.'/lib.php');
+
+
 class FlareSolverrExtension extends Minz_Extension
 {
     public function init()
     {
         $this->registerTranslates();
         $this->registerHook('post_update', array($this, 'postUpdateHook'));
+		$this->registerHook('api_misc', array($this, 'callApi'));
     }
 
     public function install()
     {
-        $filename = 'cloudsolver.php';
-        $file_source = join_path($this->getPath(), $filename);
-        $path_destination = join_path(PUBLIC_PATH, 'api');
-        $file_destination = join_path($path_destination, $filename);
 
-        if (!is_writable($path_destination)) {
-            return 'server cannot write in ' . $path_destination;
-        }
+		// for old versions of freshrss we need to inject the cloudsolver.php file
 
-        if (file_exists($file_destination)) {
-            if (!unlink($file_destination)) {
-                return 'API file seems already existing but cannot be removed';
-            }
-        }
+		if(version_compare(FRESHRSS_VERSION, "1.26.4","<")){
+			$filename = 'cloudsolver.php';
+			$file_source = join_path($this->getPath(), $filename);
+			$path_destination = join_path(PUBLIC_PATH, 'api');
+			$file_destination = join_path($path_destination, $filename);
 
-        if (!file_exists($file_source)) {
-            return 'API file seems not existing in this extension. Try to download it again.';
-        }
+			if (!is_writable($path_destination)) {
+				return 'server cannot write in ' . $path_destination;
+			}
 
-        if (!copy($file_source, $file_destination)) {
-            return 'the API file has failed during installation.';
-        }
+			if (file_exists($file_destination)) {
+				if (!unlink($file_destination)) {
+					return 'API file seems already existing but cannot be removed';
+				}
+			}
+
+			if (!file_exists($file_source)) {
+				return 'API file seems not existing in this extension. Try to download it again.';
+			}
+
+			if (!copy($file_source, $file_destination)) {
+				return 'the API file has failed during installation.';
+			}
+		}
 
         return true;
     }
@@ -48,6 +57,10 @@ class FlareSolverrExtension extends Minz_Extension
         return true;
     }
 
+	public function callApi( ) {
+		run_flaresolverr_extension();
+	}
+
     public function postUpdateHook()
     {
         $res = $this->install();
@@ -59,30 +72,46 @@ class FlareSolverrExtension extends Minz_Extension
 
     public function getPluginEndpoint()
     {
-    	return Minz_Url::display('/api/cloudsolver.php', 'html', true) . '?feed=';
+
+		if(version_compare(FRESHRSS_VERSION, "1.26.3",">")){
+			return Minz_Url::display('/api/misc.php/'.$this->getName().'?feed=', 'html', true);
+		}else{
+			return Minz_Url::display('/api/cloudsolver.php?feed=', 'html', true);
+
+		}
+
     }
 
     public function handleConfigureAction()
     {
+
+		$conf = FreshRSS_Context::systemConf();
+
+
         if (Minz_Request::isPost()) {
-            FreshRSS_Context::$system_conf->flaresolver_url = Minz_Request::param('flaresolver_url', "");
-            FreshRSS_Context::$system_conf->flaresolver_maxTimeout = Minz_Request::param("flaresolver_maxTimeout", "");
-            FreshRSS_Context::$system_conf->save();
+			$conf = FreshRSS_Context::systemConf();
+            $conf->flaresolver_url = Minz_Request::paramString('flaresolver_url', "");
+            $conf->flaresolver_maxTimeout = Minz_Request::paramInt("flaresolver_maxTimeout", "");
+            $conf->save();
         }
     }
 
     public function getFlaresolverUrl()
     {
-        if (FreshRSS_Context::$system_conf->flaresolver_url !== null)
-            return FreshRSS_Context::$system_conf->flaresolver_url;
 
-        return true;
+		if (FreshRSS_Context::systemConf()->hasParam('flaresolver_url')) {
+			return FreshRSS_Context::systemConf()->flaresolver_url;
+		}
+        return "";
     }
 
-    public function getMaxTimeout(){
-        if (FreshRSS_Context::$system_conf->flaresolver_maxTimeout !== null)
-            return intval(FreshRSS_Context::$system_conf->flaresolver_maxTimeout);
 
-        return 60000;
+
+    public function getMaxTimeout() {
+
+		if ( FreshRSS_Context::systemConf()->hasParam('flaresolver_maxTimeout') ){
+			return intval(FreshRSS_Context::systemConf()->flaresolver_maxTimeout);
+		}
+		return 60000;
     }
 }
